Prototypes: Replace "applications"
The term "applications" did not make sense, since applications aren't
covered by the copyleft rule regardless.  The longer approach would have
been to substitute "software that invokes this software's
functionality", paralleling the Copyleft section.  But that's long,
awkward to read, and tightly couples Prototypes to Copyleft.  The more
general "other software" is easier to read.  It also leaves it up to
the Copyleft section what software is covered.